directory structure components moved
